Business Partner Phone Tab Reference DDLs

+++ Cares.Web/Models/Area.cs
@@ -9,7 +9,7 @@
         /// <summary>
         /// Area ID
         /// </summary>
-        public int AreaId { get; set; }
+        public short AreaId { get; set; }
         /// <summary>
         /// Area Code
         /// </summary>
@@ -25,7 +25,7 @@
         /// <summary>
         /// City ID
         /// </summary>
-        public int CityId { get; set; }
+        public short CityId { get; set; }
         #endregion
     }
 }
